[FREELDR] Implement SYSTEM hive recovery at bootloader level & use CmCheckRegistry for registry validation
Validate the SYSTEM hive with CmCheckRegistry and purge volatile data with the same function when initializing a hive descriptor for SYSTEM.
Also implement SYSTEM recovery code that takes use of SYSTEM log in case something is fishy with the hive.

If FreeLdr repairs the hive, it will mark it with HBOOT_BOOT_RECOVERED on BootRecover field of the header. All the recovered data that is present as dirty in memory will have to be flushed by the kernel once it is in charge of the system.

=== NOTE ===

The implemented code in FreeLdr made the bootloader grew up in binary size and that prevents the booting of x64 ROS because of a check failure against the PE image of FreeLdr. As a workaround most of
the implemented code is disabled for compilation in AMD64 for now, until a real fix comes up.